Renaming for clarity and consistency
UnAuthorizedException -> UnauthorizedSafehillHttpException
SafehillApiImpl -> RemoteServer
ServerEnvironment -> RemoteServerEnvironment
ServerProxyInterface -> ServerProxy
ServerProxy -> ServerProxyImpl
