--- src/main/kotlin/com/safehill/kclient/network/ServerProxyInterface.kt
@@ -1,8 +1,0 @@
-package com.safehill.kclient.network
-
-import com.safehill.kclient.models.users.ServerUser
-
-interface ServerProxyInterface : SafehillApi {
-    suspend fun getAllLocalUsers(): List<ServerUser>
-}
-
