+++ src/main/kotlin/com/safehill/kclient/network/exceptions/UnauthorizedSafehillHttpException.kt
@@ -1,0 +1,8 @@
+package com.safehill.kclient.network.exceptions
+
+import com.safehill.kclient.network.exceptions.SafehillHttpException
+
+val UnauthorizedSafehillHttpException = SafehillHttpException(
+    statusCode = 401,
+    message = "unauthorized",
+)
