+++ src/main/kotlin/com/safehill/kclient/network/remote/RemoteServerEnvironment.kt
@@ -1,0 +1,5 @@
+package com.safehill.kclient.network.remote
+
+enum class RemoteServerEnvironment {
+    Production, Development
+}
