+++ src/test/kotlin/com/safehill/kclient/api/RemoteServerTests.kt
@@ -8,9 +8,9 @@ import com.safehill.kclient.models.assets.EncryptedAssetImpl
 import com.safehill.kclient.models.assets.EncryptedAssetVersionImpl
 import com.safehill.kclient.models.users.LocalUser
 import com.safehill.kclient.models.users.ServerUser
-import com.safehill.kclient.network.SafehillApiImpl
-import com.safehill.kclient.network.SafehillHttpException
-import com.safehill.kclient.network.SafehillHttpStatusCode
+import com.safehill.kclient.network.remote.RemoteServer
+import com.safehill.kclient.network.exceptions.SafehillHttpException
+import com.safehill.kclient.network.remote.SafehillHttpStatusCode
 import com.safehill.kcrypto.models.SafehillKeyPair
 import com.safehill.kcrypto.models.LocalCryptoUser
 import kotlinx.coroutines.CoroutineScope
@@ -21,7 +21,7 @@ import java.util.Base64
 import java.util.Date
 import kotlin.random.Random
 
-class SafehillApiImplTests {
+class RemoteServerTests {
 
     private suspend fun createUserOnServer(
         coroutineScope: CoroutineScope,
@@ -41,7 +41,7 @@ class SafehillApiImplTests {
         var error: Exception? = null
         val createJob = coroutineScope.launch {
             try {
-                serverUser = SafehillApiImpl(localUser).createUser(newUserName)
+                serverUser = RemoteServer(localUser).createUser(newUserName)
             } catch (err: Exception) {
                 error = err
             }
@@ -67,7 +67,7 @@ class SafehillApiImplTests {
 
         val authJob = coroutineScope.launch {
             try {
-                authResponse = SafehillApiImpl(localUser).signIn()
+                authResponse = RemoteServer(localUser).signIn()
             } catch (err: Exception) {
                 error = err
             }
@@ -91,7 +91,7 @@ class SafehillApiImplTests {
 
         val deleteJob = coroutineScope.launch {
             try {
-                SafehillApiImpl(localUser).deleteAccount()
+                RemoteServer(localUser).deleteAccount()
             } catch (err: Exception) {
                 error = err
             }
@@ -113,7 +113,7 @@ class SafehillApiImplTests {
 
         val deleteJob = coroutineScope.launch {
             try {
-                SafehillApiImpl(localUser).deleteAssets(assets.map { it.globalIdentifier })
+                RemoteServer(localUser).deleteAssets(assets.map { it.globalIdentifier })
             } catch (err: Exception) {
                 error = err
             }
@@ -135,7 +135,7 @@ class SafehillApiImplTests {
             var error: Exception? = null
             var getJob = launch {
                 try {
-                    val users = SafehillApiImpl(user).getUsers(listOf(user.identifier))
+                    val users = RemoteServer(user).getUsers(listOf(user.identifier))
                     assert(users.isNotEmpty())
                     val retrievedUser = users[0]
                     assert(retrievedUser.identifier == user.identifier)
@@ -166,7 +166,7 @@ class SafehillApiImplTests {
 
             val updateJob = launch {
                 try {
-                    val updatedUser = SafehillApiImpl(user).updateUser(
+                    val updatedUser = RemoteServer(user).updateUser(
                         name = null,
                         phoneNumber = newPhoneNumber,
                         email = null
@@ -187,7 +187,7 @@ class SafehillApiImplTests {
 
             val authJob = launch {
                 try {
-                    val authResponse = SafehillApiImpl(user).signIn()
+                    val authResponse = RemoteServer(user).signIn()
                     assert(authResponse.metadata.isPhoneNumberVerified)
                 } catch (err: Exception) {
                     error = err
@@ -203,7 +203,7 @@ class SafehillApiImplTests {
 
             getJob = launch {
                 try {
-                    val users = SafehillApiImpl(user).getUsers(listOf(user.identifier))
+                    val users = RemoteServer(user).getUsers(listOf(user.identifier))
                     assert(users.isNotEmpty())
                     assert(users.count() == 1)
                     val retrievedUser = users[0]
@@ -256,7 +256,7 @@ class SafehillApiImplTests {
             val user = createUserOnServer(this)
             authenticateUser(this, user)
 
-            val api = SafehillApiImpl(user)
+            val api = RemoteServer(user)
 
             var createdAsset: com.safehill.kclient.models.dtos.AssetOutputDTO? = null
             var error: Exception? = null
@@ -307,7 +307,7 @@ class SafehillApiImplTests {
     fun testUnauthorizedGetUsers() {
         val cryptoUser = LocalCryptoUser()
         val localUser = LocalUser(cryptoUser)
-        val api = SafehillApiImpl(localUser)
+        val api = RemoteServer(localUser)
 
         runBlocking {
             try {
@@ -335,7 +335,7 @@ class SafehillApiImplTests {
     @Test
     fun testAuthenticateNonExistingUser() {
         val localUser = LocalUser(LocalCryptoUser())
-        val api = SafehillApiImpl(localUser)
+        val api = RemoteServer(localUser)
 
         runBlocking {
             try {
@@ -353,7 +353,7 @@ class SafehillApiImplTests {
             authenticateUser(this, user)
 
             var error: Exception? = null
-            val api = SafehillApiImpl(user)
+            val api = RemoteServer(user)
 
             try {
                 api.sendCodeToUser(1, 4151234567, "12345", SendCodeToUserRequestDTO.Medium.SMS)
